index.js: handle /dev/null on windows
On windows the /dev/null device is not existing, treat it as a null
device in this case.

Assume /dev/null generates an empty string in case of source file, and
assume /dev/null is a null sync in case of destination file.

This fixes an use case when the patch is generated e.g. with:
diff -u /dev/null test > add.patch

In case of a null output devices, e.g. with:
diff -u test /dev/null > drop.patch

no output is written (as we are implicitly assuming to write to the
/dev/null file).